+++ tools4j-sequences/src/main/java/com/github/panpf/tools4j/sequences/Sequencex.java
@@ -2086,6 +2086,183 @@ public class Sequencex {
     }
 
 
+    /**
+     * Returns the largest value among all values produced by [selector] function
+     * applied to each element in the sequence.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T> Double maxOfByDouble(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Double> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        double maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            double v = selector.transform(iterator.next());
+            maxValue = Math.max(maxValue, v);
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value among all values produced by [selector] function
+     * applied to each element in the sequence.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T> Float maxOfByFloat(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Float> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        float maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            float v = selector.transform(iterator.next());
+            maxValue = Math.max(maxValue, v);
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value among all values produced by [selector] function
+     * applied to each element in the sequence.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T, R extends Comparable<R>> R maxOf(@Nullable Sequence<T> sequence, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        R maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (maxValue.compareTo(v) < 0) {
+                maxValue = v;
+            }
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value among all values produced by [selector] function
+     * applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T> Double maxOfByDoubleOrNull(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Double> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        double maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            double v = selector.transform(iterator.next());
+            maxValue = Math.max(maxValue, v);
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value among all values produced by [selector] function
+     * applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T> Float maxOfByFloatOrNull(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Float> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        float maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            float v = selector.transform(iterator.next());
+            maxValue = Math.max(maxValue, v);
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value among all values produced by [selector] function
+     * applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T, R extends Comparable<R>> R maxOfOrNull(@Nullable Sequence<T> sequence, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        R maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (maxValue.compareTo(v) < 0) {
+                maxValue = v;
+            }
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value according to the provided [comparator]
+     * among all values produced by [selector] function applied to each element in the sequence.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T, R> R maxOfWith(@Nullable Sequence<T> sequence, @NotNull Comparator<R> comparator, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        R maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (comparator.compare(maxValue, v) < 0) {
+                maxValue = v;
+            }
+        }
+        return maxValue;
+    }
+
+    /**
+     * Returns the largest value according to the provided [comparator]
+     * among all values produced by [selector] function applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T, R> R maxOfWithOrNull(@Nullable Sequence<T> sequence, @NotNull Comparator<R> comparator, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        R maxValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (comparator.compare(maxValue, v) < 0) {
+                maxValue = v;
+            }
+        }
+        return maxValue;
+    }
+
+
     /* ******************************************* min ******************************************* */
 
 
@@ -2190,6 +2367,183 @@ public class Sequencex {
             if (comparator.compare(min, e) > 0) min = e;
         }
         return min;
+    }
+
+
+    /**
+     * Returns the smallest value among all values produced by [selector] function
+     * applied to each element in the sequence.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T> Double minOfByDouble(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Double> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        double minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            double v = selector.transform(iterator.next());
+            minValue = Math.min(minValue, v);
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value among all values produced by [selector] function
+     * applied to each element in the sequence.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T> Float minOfByFloat(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Float> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        float minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            float v = selector.transform(iterator.next());
+            minValue = Math.min(minValue, v);
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value among all values produced by [selector] function
+     * applied to each element in the sequence.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T, R extends Comparable<R>> R minOf(@Nullable Sequence<T> sequence, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        R minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (minValue.compareTo(v) > 0) {
+                minValue = v;
+            }
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value among all values produced by [selector] function
+     * applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T> Double minOfByDoubleOrNull(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Double> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        double minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            double v = selector.transform(iterator.next());
+            minValue = Math.min(minValue, v);
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value among all values produced by [selector] function
+     * applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * If any of values produced by [selector] function is `NaN`, the returned result is `NaN`.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T> Float minOfByFloatOrNull(@Nullable Sequence<T> sequence, @NotNull Transformer<T, Float> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        float minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            float v = selector.transform(iterator.next());
+            minValue = Math.min(minValue, v);
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value among all values produced by [selector] function
+     * applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T, R extends Comparable<R>> R minOfOrNull(@Nullable Sequence<T> sequence, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        R minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (minValue.compareTo(v) > 0) {
+                minValue = v;
+            }
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value according to the provided [comparator]
+     * among all values produced by [selector] function applied to each element in the sequence.
+     *
+     * @throws NoSuchElementException if the sequence is empty.
+     *                                <p>
+     *                                The operation is _terminal_.
+     */
+    @NotNull
+    public static <T, R> R minOfWith(@Nullable Sequence<T> sequence, @NotNull Comparator<R> comparator, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) throw new NoSuchElementException();
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) throw new NoSuchElementException();
+        R minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (comparator.compare(minValue, v) > 0) {
+                minValue = v;
+            }
+        }
+        return minValue;
+    }
+
+    /**
+     * Returns the smallest value according to the provided [comparator]
+     * among all values produced by [selector] function applied to each element in the sequence or `null` if there are no elements.
+     * <p>
+     * The operation is _terminal_.
+     */
+    @Nullable
+    public static <T, R> R minOfWithOrNull(@Nullable Sequence<T> sequence, @NotNull Comparator<R> comparator, @NotNull Transformer<T, R> selector) {
+        if (sequence == null) return null;
+        Iterator<T> iterator = sequence.iterator();
+        if (!iterator.hasNext()) return null;
+        R minValue = selector.transform(iterator.next());
+        while (iterator.hasNext()) {
+            R v = selector.transform(iterator.next());
+            if (comparator.compare(minValue, v) > 0) {
+                minValue = v;
+            }
+        }
+        return minValue;
     }