+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * new: Sequencex add flatMapOfIterable, flatMapOfIterableTo, flatMapIndexed, flatMapIndexedOfIterable, flatMapIndexedOfIterableTo, flatMapIndexedTo method
 * new: Collectionx add flatMapIndexed, flatMapIndexedTo method
 * change: Sequencex max\* and min\* method rename to max\*OrNull and min\*OrNull
+* new: Sequencex added maxOf\* and maxOf\* method
 
 ## v1.0.0-beta02
 improve: Improve code